Adding some methods for decimal helper and create account dto.

+++ src/main/java/com/gbm/challenge/gbmchallenge/model/request/CreateAccountDto.java
@@ -3,11 +3,9 @@ package com.gbm.challenge.gbmchallenge.model.request;
 import lombok.Getter;
 import lombok.NonNull;
 
-import java.math.BigDecimal;
-
 @Getter
 public class CreateAccountDto {
 
-    @NonNull private BigDecimal cash;
+    @NonNull private Double cash;
 
 }
